AppsEndpointsTests
